+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Changelog with backdated changes (#51)
 * List filtering (#54)
 * Show installations on native objects (#49)
+* SPDX license expressions (#47)
 
 ### Changed